Record referrer on bundle (#8)
In order to track where our transaction flow is coming from we receive an optional referrer tag in the header of the request, that we now will add into Dune.

Since we a re adding a column, this will require indexing on their end. It's my understanding that this won't lead to any data loss though, so it's probably cleaner than adding it to the transaction json (where it would be duplicated and not semantically it isn't really part of a transaction).